Fix radius_knn
- In `radius_knn`, for each original data point, we get the nearest
  neighbors within the radius r (r being the distance between the
  original data point and its avatar).
- We then compute the number of avatars and the number of records within
  that radius.

The issue is that when we compute the number of avatars, we don't
exclude the avatar generated by this data point. In some cases, we thus
report an avatar count of 1 even though there is no avatar separating
the two points.

The key code change is here:

```python
avatars_count = sum(1 for k in inside_radius_list[1][0] if k > (N - 1) and k != i + N)
records_count = sum(1 for k in inside_radius_list[1][0] if k < (N - 1))
```